Update trips-simulation.R
Please, check if this final section for calculating impact using distance is correct. The idea is to have a measure of the changes in km. My doubt is if the results here indicates the change in the universe of the trips or only the sample used to generate the scenarios.